requirements.txt
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
coverage==7.2.5
exceptiongroup==1.1.1
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
importlib-metadata==6.5.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
pluggy==1.0.0
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg-pool==3.1.7
PyJWT==2.7.0
pytest==7.3.1
pytest-cov==4.0.0
pytest-mock==3.10.0
python-dotenv==1.0.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
Werkzeug==2.2.3
zipp==3.15.0
